fix(ci): clear module export migration guardrails
